reintegration of ulong branch, plus some other changes accidentally checked in to that branch: ulong changes: - added stdint.h header for msvc because it isn't included by default, adds (u)int{8,16,32,64}_t types - changed "unsigned long" declarations in kernel to uint32_t when the 32-bit size was assumed and required, mainly hash values/ids and the functions that produce and use them - demonstrated increase in performance with 64-bit builds with learning on - updated retesave_{one_byte,two_bytes,four_bytes} functions to use the standard uint{8,16,32}_t types - changed "unsigned long long" to uint64_t misc changes: - rewrote a few inline functions so they better represent the macros that they replaced, some were introducing extra variables and casts that wouldn't be introduced if the macros were used, removed unnecessary casts - deallocate_instantiation algorithm changed to prevent a stack overflow triggered by retracting an instantiation with a huge chain of backtrace instantiations, see more comments in deallocate_instantiation, basically flattened out a recursive loop - added test to verify deallocate_instantiation change

+++ SoarSuite/Core/SoarKernel/include/chunk.h
@@ -37,8 +37,8 @@ typedef struct chunk_cond_struct {
   /* dll of cond's in this particular hash bucket for this set */
   struct chunk_cond_struct *next_in_bucket, *prev_in_bucket; 
 
-  unsigned long hash_value;             /* equals hash_condition(cond) */
-  unsigned long compressed_hash_value;  /* above, compressed to a few bits */
+  uint32_t hash_value;             /* equals hash_condition(cond) */
+  uint32_t compressed_hash_value;  /* above, compressed to a few bits */
 } chunk_cond;
